Refactor filter table
Rather than maintain a huge list, in python code, of filter names and
comments, lets just use the filter comments shipped with FSPS itself.
The filter names are unique to python-fsps, and stored separately in a
filter_keys.txt. The two data sets are merged and turned into a
dictionary of Filter instances with the load_filter_dict() function.

Hopefully this will make the filter table easier to maintain