get_mags() returns mags in same order as bands arg
Before, if the user specified the bands argument to
StellarPopulation.get_mags() then the subset of magnitudes would be
returned in the order of FSPS filter indices. The user would probably
expect the mags to match the order of the bands list since the FSPS band
order is rather arbitrary. This commit satisfies that.