Edited dust1_index docstring to reflect usage
Previously, the docstring (and therefore documentation on the
website) stated that dust1_index is only used when dust_type=0.
This is not the case, dust1_index is used for all dust types,
as can be seen in the Fortran FSPS documentation. I have also
verified this in in python-fsps. This commit changes the docstring
to correct this.